+++ src/operations/padOperation.ts
@@ -11,6 +11,7 @@
 import { execSync } from "child_process";
 
 import {
+  execSIPSCommandOnAVIF,
   execSIPSCommandOnSVG,
   execSIPSCommandOnWebP,
   getDestinationPaths,
@@ -41,16 +42,24 @@ export default async function pad(sourcePaths: string[], padding: number, color:
       resultPaths.push(
         await execSIPSCommandOnWebP(
           `sips --padToHeightWidth ${oldHeight + padding} ${oldWidth + padding} --padColor ${color}`,
-          imagePath
-        )
+          imagePath,
+        ),
       );
     } else if (imagePath.toLowerCase().endsWith(".svg")) {
       // Convert to PNG, apply padding, then restore to SVG
       resultPaths.push(
         await execSIPSCommandOnSVG(
           `sips --padToHeightWidth ${oldHeight + padding} ${oldWidth + padding} --padColor ${color}`,
-          imagePath
-        )
+          imagePath,
+        ),
+      );
+    } else if (imagePath.toLowerCase().endsWith(".avif")) {
+      // Convert to PNG, apply padding, then restore to AVIF
+      resultPaths.push(
+        await execSIPSCommandOnAVIF(
+          `sips --padToHeightWidth ${oldHeight + padding} ${oldWidth + padding} --padColor ${color}`,
+          imagePath,
+        ),
       );
     } else {
       // Image is not a special format, so pad using SIPS
@@ -60,7 +69,7 @@ export default async function pad(sourcePaths: string[], padding: number, color:
       execSync(
         `sips --padToHeightWidth ${oldHeight + padding} ${
           oldWidth + padding
-        } --padColor ${color} -o "${newPath}" "${imagePath}"`
+        } --padColor ${color} -o "${newPath}" "${imagePath}"`,
       );
     }
   }